CORDA-3495: Incorrect links in the docs
Several links to released Corda binaries were found to be incorrect. They had -SNAPSHOT in their URLs.

This is because the corda_version property is used when generating the URLs. As the corda_version property contains -SNAPSHOT, this will reoccur.

Therefore, a new property was introduced - stable_corda_version. This is intended to represent the most recently released stable version of Corda, 4.3 at the time of writing.