CORDA-5725: Optimize serialization by caching and reusing the serialized schema
Serialization encodes both the data and its schema. The schema takes up to 90% of the space and an equal amount of CPU time to serialize. This can lead to a significant bottleneck because of the use of serialization not only for the serialization itself, but also for the computation of merkle trees and transaction ids. A cache for the serialized schema is introduced to avoid this performance penalty:

- the schema gets a unique identifier (used as cache key)
- serialization is split into data, schema and transformation serialization and then combined together with the necessary Corda/AMQP envelop

The same technique could be applied for deserialization and DBTransactionStorage in subsequent commits.