ENT-4601 Return `CompletableFuture` from `FlowExternalFuture`
Replace the `CordaFuture` returned from `FlowExternalFuture.execute` with a
`CompletableFuture`. This is to allow developers to use the functions that
Java provides out of the box for handling the creation and completion of
futures.

Part of the decision to do this, is the fact that all of the `CordaFuture`
functions are part of the internal API.

Using `CompletableFuture` is easier from a users perspective as they would
likely create one anyway, and then wrap/convert it to a `CordaFuture`.