added `queryLazy`public API
### Problem
the vault query API currently requires the client to handle pagination explicitly and, if he doesn't (which is easy since `net.corda.core.node.services.VaultService#queryBy(java.lang.Class<? extends T>, net.corda.core.node.services.vault.QueryCriteria)` is publicly available and only requires an instance of `QueryCriteria` as an input), it throws an exception
whenever the number of results exceeds `DEFAULT_PAGE_SIZE` (which is currently 200). This makes is quite nasty because the code that naively use that method will easily pass unit tests, that commonly makes use of limited size dataset, only to fail as soon as the vault starts to contain a more realistic amount of data.

### Currently implemented solution
To overcome this limitation I decided to add the `queryLazy` method to both `net.corda.core.node.services.VaultService` and `net.corda.core.messaging.CordaRPCOps`, this will return a `kotlin.collections.Iterator` in Kotlin that can be easily converter into a `Sequence` with the `asSequence` method, in Java it returns a `java.util.Iterator` that can be used to create a `java.lang.Iterable` on the fly (which in turn can be used to create a `java.util.stream.Stream` using `java.util.stream.StreamSupport#stream(java.util.Spliterator<T>, boolean)`).

The iterator will perform an immediate vault query at the time of its creation and then it will lazily query the vault as soon as a new page is required while the client is traversing the result list.

The size of the page can be specified as the last function parameter to meet client memory/latency requirements, if that parameter is omitted, it will default to `DEFAULT_PAGE_SIZE`

### Current limitations
The inline function `net.corda.core.node.services.VaultServiceKt#queryLazy` needs to be **public** to be usable whilce it would be better to have it only internally exposed. Being an *inline* function, it is only callable from Kotlin though.
I wrote it because it is easier to use from Kotlin (you don't have to pass the class of the generic parameter as an argument) and it's marginally more efficient (avoiding the nested function call).