CORDA-3596 Record flow metadata
For all types of flow starts, flow metadata is recorded. The type of
flow start determines what data can be recorded.

Metadata is recorded in the following classes:

    * RPC = `CordaRPCOpsImpl`
    * SERVICE = `AppServiceHubImpl`
    * INITIATED = `SingleThreadedStateMachineManager`
    * SCHEDULED = `NodeSchedulerService`

* `FlowMetadataRecorder` delegates to `CheckpointStorage` to record the
metadata of a flow. This class is called from all points where a new
flow is started. A new database transaction is opened to save the
metadata as this function is called outside the context of a flow and
therefore an existing database transaction.

* `CheckpointStorage` has a new function, `addMetadata` and is
implemented in `DBCheckpointStorage`. For now, a new class has been
added inside of the internal (`FlowMetadata`) as a wrapper object that
can be passed to `addMetadata`. This stops the entity from leaking out
of `DBCheckpointStorage`.

* Various classes have had their constructors changed to allow the
recording classes to be passed into them.