ENT-4982 Flush vault changes to underlying session
The `CashSelectionTest` highlighted a change in functionality due to
previous changes.

Saved transactions and states were not being flushed to the underlying
session inside of `NodeVaultService.notifyAll`. Every time a batch of
transaction updates are made to the database they are now also flushed.

This allows other operations inside the same database transaction to
observe the updates to the database that the previous calls made.