ENT-4982 Closing a database transaction does not roll it back
## Description

Whether a database transaction is rolled back when closing its
connection/session depends on the database and its JDBC driver
(according to the internet anyway).

When an uncaught error propagates through the state machine, the current
database transaction is rolled back
(via `ActionExecutorImpl.executeRollbackTransaction`). How it did so,
was by just closing the transaction.

This did not seem to roll back the transaction, instead it was
committing uncommitted changes that were part of the current
transaction's session. This lead to data being persisted to the database
even though the flow failed. Furthermore, if the flow was to retry,
constraint violations could arise due to data that should not have been
committed actually residing in the database.

## Solution

The database transaction is now explicitly rolled back from two points:

- `ActionExecutorImpl.executeRollbackTransaction`
- `TransitionExecutorImpl.executeTransition`

All other instances of `DatabaseTransaction.close` were already rolling
back the transaction if an error occurred.

By explicitly rolling back, we no longer depend on the JDBC driver's
implementation to decide the behaviour.