ENT-4982 Unrestrict functions
Allow a transaction marked for rollback to allow update statements to be
executed on it, even though none of the updates will actually go
through. Allow `EntityManager.transaction` to be called but restrict the
functions that can be called on it. This is to allow flows to control
what happens themselves, rather than enforcing it from Corda's
perspective.