CORDA-3291 Keep track of future in `StateMachineState`
Instead of relying on a map containing all the sleeping flows and
their futures, keep the future on the `StateMachineState` and manage it
directly there.

This removes the need for concurrency control around flow sleeps as the
code path does not need to touch any concurrent data structures.

To achieve this:

- `StateMachineState.future` added as a `var`
- `StateMachineState.isWaitingForFuture` __changed__ to a `var`
- When the `ScheduledFuture` is created to wake up the flow the passed
in `StateMachineState` has its `future` value changed
- When resumed `future` and `isWaitingForFuture` are set to `null` and
`false` respectively
- When cancelling a sleeping flow, the `future` is cancelled and nulled
out and `isWaitingForFuture` is set to false (this only happens if the
flow was actually sleeping)