INFRA-438 Handle observers not having error handling
When the RPC client connection is closed, it notifies observers using onError(),
which may not be the correct approach (TBD) but changing this is a much more invasive change.
Where observers do not subscribe to error notifications, this is reflected to the calling
client by an exception thrown.

This change catches that exception and lots it as debug rather an error level.