CORDA-3777 Use count instead of flag
To make the reloading work for idempotent flows and for flows that skip
saving their checkpoints, a count is maintained for what suspension
point was reached and reloaded from.

This means idempotent flows can reload from the start and not reload
again until reaching a new suspension point.

Flows that skip checkpoints can reload from a previously saved
checkpoint (or from the initial checkpoint) and will continue reloading
on reaching the next new suspension point (not the suspension point that
it skipped saving).